apps/agent-core/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pyyaml>=6.0.1
httpx>=0.28.0
pytest>=8.0.0
